+++ dev/stan.clj
@@ -8,14 +8,16 @@
 
 (comment
 
-  (apps/make-docker-artifact {:main-namespace    'codesmith.anvil.artifacts
-                              :lib-name          'anvic
-                              :aliases           []
-                              :version           "1.0.0"
-                              :java-version      :openjdk/jdk17
-                              :docker-base-image "other-base-image"
-                              :basis-opts        {:user :standard}
-                              :aot?              false})
+  (apps/make-docker-artifact {:main-namespace 'codesmith.anvil.artifacts
+                              :lib-name       'anvic
+                              :aliases        []
+                              :version        "1.0.0"
+                              :java-runtime   {:version           :java17
+                                               :type              :jre
+                                               :modules-profile   :anvil
+                                               :extra-modules     #{"java.se"}}
+                              :basis-opts     {:user :standard}
+                              :aot?           false})
 
   (keys
     (b/create-basis {:user :standard}))