+++ test/helloworld/src/hello.clj
@@ -1,4 +1,5 @@
 (ns hello)
 
 (defn -main [& args]
-  (println "Hello World!"))
+  (println "Hello World!")
+  (println (slurp "/app/version.edn")))